Circuit-break scd41 crashes and use low power measurement.
Seems that sometime the sdc41 fails, so is needed to escape
from the loop otherwise the main loop will be blocked forever.
Failure seems related to power issues, so we switch to
low power measurement, but still from time to time fails,
propably because main PSU cannot cope with 200mAh spikes during the read.
Need to check if is a cable issues or we need to replace the sensor,
maybe with the SCD30 which is a bit more costly but consumes less at peak
(and is more precise)